// src/models/proficiency.ts
export enum Proficiency {
	None = 'None',
	Any = 'Any proficiency',
	All = 'All proficiencies',
	MilitaryWeapons = 'Military weapons',
	LargeWeapons = 'Large weapons',
	PairedWeapons = 'Paired weapons',
	RangedWeapons = 'Ranged weapons',
	PowderWeapons = 'Powder weapons',
	Implements = 'Implements',
	LightArmor = 'Light armor',
	HeavyArmor = 'Heavy armor',
	Shields = 'Shields'
}